chore: 额外检查

+++ biliup/plugins/youtube.py
@@ -18,21 +18,24 @@ from ..engine.download import DownloadBase
 VALID_URL_BASE = r'https?://(?:(?:www|m)\.)?youtube\.com/(?P<id>(?!.*?/live$).*?)\??(.*?)'
 VALID_URL_LIVE = r'https?://(?:(?:www|m)\.)?youtube\.com/(?P<id>.*?)/live'
 
+# proxy = "http://127.0.0.1:7890"
+proxy = None
+
 @Plugin.download(regexp=VALID_URL_LIVE)
 class YoutubeLive(DownloadBase):
     def __init__(self, fname, url, suffix='flv'):
         super().__init__(fname, url, suffix)
-        self.ytb_cookie = config.get('user', {}).get('youtube_cookie')
+        self.youtube_cookie = config.get('user', {}).get('youtube_cookie')
         self.cache_dir = f"./cache/{self.__class__.__name__}/{self.fname}"
         self.__webpage_url = None
 
     async def acheck_stream(self, is_check=False):
         ydl_opts = {
             'download_archive': f"{self.cache_dir}/archive.txt",
-            'cookiefile': self.ytb_cookie,
+            'cookiefile': self.youtube_cookie,
             'ignoreerrors': True,
             'extractor_retries': 0,
-            # 'proxy': "socks5://127.0.0.1:7890",
+            'proxy': proxy,
         }
         try:
             # video_id = self.get_video_id_from_archive(f"{self.cache_dir}/archive.txt")
@@ -72,7 +75,7 @@ class YoutubeLive(DownloadBase):
 
     def download(self):
         # 归档后封面不允许下载，需提前下载
-        self.use_live_cover = True
+        # self.use_live_cover = True
         if self.use_live_cover:
             # 后台下载封面
             cover_thread = threading.Thread(
@@ -87,9 +90,13 @@ class YoutubeLive(DownloadBase):
 
         # self.downloader = 'ytarchive'
 
+        # 检查 cache_dir 属性是否存在，如果不存在则创建
+        if not hasattr(self, 'cache_dir'):
+            self.cache_dir = f"./cache/{self.__class__.__name__}/{self.fname}"
+
         # stream-gears 和 streamlink(sync-downloader) 交给父类下载器来支持分段
         if self.downloader in ['stream-gears', 'streamlink', 'sync-downloader']:
-            if self.downloader != 'stream-gears':
+            if self.downloader != 'stream-gears' and self.__webpage_url:
                 # 让 streamlink 自行提取
                 self.raw_stream_url = self.__webpage_url
                 # pass
@@ -100,17 +107,17 @@ class YoutubeLive(DownloadBase):
             'temporary-dir': self.cache_dir,
             'threads': 3,
             'output': f"{filename}.{self.suffix}",
-            # 'proxy': "http://127.0.0.1:7890",
-            'cookies': self.ytb_cookie,
+            'proxy': proxy,
+            'cookies': self.youtube_cookie,
             'add-metadata': True,
             'newline': True,
         }
         ydl_opts = {
             'outtmpl': f"{self.cache_dir}/{filename}.%(ext)s",
-            'cookiefile': self.ytb_cookie,
+            'cookiefile': self.youtube_cookie,
             'break_on_reject': True,
             'format': 'best',
-            # 'proxy': "http://127.0.0.1:7890",
+            'proxy': proxy,
         }
 
         if self.downloader == 'ytarchive':
@@ -149,8 +156,9 @@ class YoutubeLive(DownloadBase):
                 with yt_dlp.YoutubeDL(ydl_opts) as ydl:
                     ydl.download([self.__webpage_url])
                     # 下载成功的情况下移动到运行目录
-                    for file in os.listdir(self.cache_dir):
-                        shutil.move(f'{self.cache_dir}/{file}', '.')
+                    if os.path.exists(self.cache_dir):
+                        for file in os.listdir(self.cache_dir):
+                            shutil.move(f'{self.cache_dir}/{file}', '.')
             except DownloadError as e:
                 if 'ffmpeg is not installed' in e.msg:
                     logger.error(f"{self.plugin_msg}: ffmpeg未安装，无法下载")
@@ -160,7 +168,8 @@ class YoutubeLive(DownloadBase):
             finally:
                 # 清理意外退出可能产生的多余文件
                 try:
-                    shutil.rmtree(self.cache_dir)
+                    if os.path.exists(self.cache_dir):
+                        shutil.rmtree(self.cache_dir)
                 except:
                     logger.error(f"{self.plugin_msg}: 清理残留文件失败 -> {self.cache_dir}")
 
@@ -193,6 +202,7 @@ class Youtube(DownloadBase):
         self.youtube_after_date = config.get('youtube_after_date')
         self.youtube_enable_download_live = config.get('youtube_enable_download_live', True)
         self.youtube_enable_download_playback = config.get('youtube_enable_download_playback', True)
+        self.is_live = False
         # 需要下载的 url
         self.download_url = None
 
@@ -202,6 +212,7 @@ class Youtube(DownloadBase):
             'cookiefile': self.youtube_cookie,
             'ignoreerrors': True,
             'extractor_retries': 0,
+            'proxy': proxy,
         }) as ydl:
             # 获取信息的时候不要过滤
             ydl_archive = copy.deepcopy(ydl.archive)
@@ -285,6 +296,7 @@ class Youtube(DownloadBase):
                     self.room_title = download_entry.get('title')
                     self.live_cover_url = download_entry.get('thumbnail')
                     self.download_url = download_entry.get('webpage_url')
+                    # self.is_live = download_entry.get('live_status') == 'is_live'
                 return True
             else:
                 return False
